+++ Security/Firewall/WsseListener.php
@@ -10,6 +10,7 @@ use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Http\Firewall\ListenerInterface;
+use Symfony\Component\Translation\TranslatorInterface;
 
 /**
  * 
@@ -33,47 +34,64 @@ class WsseListener extends LoggerHelper implements ListenerInterface
 	 */
 	protected $errorFactory;
 	
-	public function __construct(TokenStorageInterface $tokenStorage, AuthenticationManagerInterface $authenticationManager, ErrorResponseFactory $errorFactory)
+	/**
+	 * @var TranslatorInterface $translator
+	 */
+	protected $translator;
+	
+	/**
+	 * @param TokenStorageInterface $tokenStorage
+	 * @param AuthenticationManagerInterface $authenticationManager
+	 * @param ErrorResponseFactory $errorFactory
+	 * @param TranslatorInterface $translator
+	 */
+	public function __construct(TokenStorageInterface $tokenStorage, AuthenticationManagerInterface $authenticationManager, ErrorResponseFactory $errorFactory, TranslatorInterface $translator)
 	{
 		$this->tokenStorage = $tokenStorage;
 		$this->authenticationManager = $authenticationManager;
 		$this->errorFactory = $errorFactory;
+		$this->translator = $translator;
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 * @see \Symfony\Component\Security\Http\Firewall\ListenerInterface::handle()
+	 */
 	public function handle(GetResponseEvent $event)
 	{
-		$matches = [];
 		$request = $event->getRequest();
 		$headers = $request->headers;
-		$failedMessage = Response::$statusTexts[Response::HTTP_FORBIDDEN];
 		
-		// Verifie que la requête comporte bien le header X-WSSE
-		if ($headers->has('x-wsse') && preg_match('#UsernameToken Username="([^"]+)", PasswordDigest="([^"]+)", Nonce="([^"]+)", Created="([^"]+)"#', $headers->get('x-wsse'), $matches)) {
-			$token = new WsseUserToken();
-			$token->setUser($matches[1]);
-			$token->digest 	= $matches[2];
-			$token->nonce 	= $matches[3];
-			$token->created = $matches[4];
+		$matches = [];
+		$credentials = $headers->get('x-wsse', '');
+		$failedMessage = Response::$statusTexts[Response::HTTP_UNAUTHORIZED];
+		
+		// Verifie que le header X-WSSE est bien définie
+		if ($credentials !== '' && preg_match('#UsernameToken Username="([^"]+)", PasswordDigest="([^"]+)", Nonce="([^"]+)", Created="([^"]+)"#', $credentials, $matches)) {
+			$preAuthToken = new WsseUserToken($matches[1], $credentials, []);
+			$preAuthToken->setAttribute('digest', $matches[2]);
+			$preAuthToken->setAttribute('nonce', $matches[3]);
+			$preAuthToken->setAttribute('created', $matches[4]);
 			
 			try {
-				$authToken = $this->authenticationManager->authenticate($token);
+				$authToken = $this->authenticationManager->authenticate($preAuthToken);
 				$this->tokenStorage->setToken($authToken);
-				
 				return;
+				
 			} catch (\Exception $e) {
-				$failedMessage = $e->getMessage();
+				$failedMessage = $e->getMessage();				
 				$this->logDebug(sprintf('Login with WS-Security failed, caused by : %s', $e->getMessage()), [
-						'username'	=> $token->getUsername(),
-						'digest'	=> $token->digest,
-						'nonce'		=> $token->nonce,
-						'created'	=> $token->created
+						'username'	=> $preAuthToken->getUsername(),
+						'digest'	=> $preAuthToken->getAttribute('digest'),
+						'nonce'		=> $preAuthToken->getAttribute('nonce'),
+						'created'	=> $preAuthToken->getAttribute('created')
 				]);
 			}
 		}
 		
 		// Deny authentication with a '401 Unauthorized' HTTP response
 		$event->setResponse($this->errorFactory->create(
-				$failedMessage, 
+				$this->translator->trans($failedMessage, [], 'OkaApiBundle'), 
 				Response::HTTP_UNAUTHORIZED, 
 				null, 
 				[],